+++ docs/changelog.rst
@@ -5,6 +5,8 @@ Changelog
 
 Changes in 0.8.2
 ================
+- DateTimeField now auto converts valid datetime isostrings into dates (#343)
+- DateTimeField now uses dateutil for parsing if available (#343)
 - Fixed Doc.objects(read_preference=X) not setting read preference (#352)
 - Django session ttl index expiry fixed (#329)
 - Fixed pickle.loads (#342)